refactor: 更新 StepIndicator 的引用路徑至 common 目錄，並刪除不再使用的元件以改善檔案組織。
