+++ src/components/common/StepIndicator.tsx
